requirements.txt
agno>=1.4.0
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
apscheduler>=3.10.4
duckduckgo-search>=6.1.0
anthropic>=0.28.0
python-dotenv>=1.0.1
rich>=13.7.1
httpx>=0.27.0
pydantic>=2.7.0